change type to float16

+++ extract_features.py
@@ -428,8 +428,8 @@ def main(_):
     with tf.gfile.Open(FLAGS.output_file, "w") as f:
         final_matrix = []
         for i, result in enumerate(estimator.predict(input_fn, yield_single_examples=True)):
-            layer_output = result["layer_output_0"]
-            # print(layer_output)
+            layer_output = result["layer_output_0"].astype(np.float16)
+            # print(layer_output.dtype)  # float16
             final_matrix.append(layer_output)
         np.save(f, np.array(final_matrix))
